// src/app/api/model/step-up-action-type.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum StepUpActionType {
    EMAIL_CHANGE = 'EMAIL_CHANGE',
    PASSWORD_CHANGE = 'PASSWORD_CHANGE'
}