// src/app/api/model/step-up-challenge-type.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum StepUpChallengeType {
    EMAIL_CODE = 'EMAIL_CODE',
    TOTP = 'TOTP'
}