// src/app/api/model/step-up-token-response.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StepUpTokenResponse { 
    expiresInSeconds?: number;
    success?: boolean;
    token?: string;
}